feat(model): Add support for non-abridged objects in Model.from_dict
This commit adds support for non-abridged objects within the Model.from_dict method.

It also cleans up the ColorRoom class by pulling out the properties/methods that are shared with the new ColorFace class.
