+++ tests/model_extend_test.py
@@ -631,6 +631,54 @@ def test_to_dict_air_walls():
     assert model_idf_str.count('ZoneMixing') == 16
 
 
+def test_from_dict_non_abridged():
+    """Test the Model from_dict method with non-abridged objects."""
+    first_floor = Room.from_box('First Floor', 10, 10, 3, origin=Point3D(0, 0, 0))
+    second_floor = Room.from_box('Second Floor', 10, 10, 3, origin=Point3D(0, 0, 3))
+    first_floor.properties.energy.program_type = office_program
+    second_floor.properties.energy.program_type = office_program
+    first_floor.properties.energy.add_default_ideal_air()
+    second_floor.properties.energy.add_default_ideal_air()
+    for face in first_floor[1:5]:
+        face.apertures_by_ratio(0.2, 0.01)
+    for face in second_floor[1:5]:
+        face.apertures_by_ratio(0.2, 0.01)
+
+    pts_1 = [Point3D(0, 0, 6), Point3D(0, 10, 6), Point3D(10, 10, 6), Point3D(10, 0, 6)]
+    pts_2 = [Point3D(0, 0, 6), Point3D(5, 0, 9), Point3D(5, 10, 9), Point3D(0, 10, 6)]
+    pts_3 = [Point3D(10, 0, 6), Point3D(10, 10, 6), Point3D(5, 10, 9), Point3D(5, 0, 9)]
+    pts_4 = [Point3D(0, 0, 6), Point3D(10, 0, 6), Point3D(5, 0, 9)]
+    pts_5 = [Point3D(10, 10, 6), Point3D(0, 10, 6), Point3D(5, 10, 9)]
+    face_1 = Face('Attic Face 1', Face3D(pts_1))
+    face_2 = Face('Attic Face 2', Face3D(pts_2))
+    face_3 = Face('Attic Face 3', Face3D(pts_3))
+    face_4 = Face('Attic Face 4', Face3D(pts_4))
+    face_5 = Face('Attic Face 5', Face3D(pts_5))
+    attic = Room('Attic', [face_1, face_2, face_3, face_4, face_5], 0.01, 1)
+
+    constr_set = ConstructionSet('Attic Construction Set')
+    polyiso = EnergyMaterial('PolyIso', 0.2, 0.03, 43, 1210, 'MediumRough')
+    roof_constr = OpaqueConstruction('Attic Roof Construction',
+                                     [roof_membrane, polyiso, wood])
+    floor_constr = OpaqueConstruction('Attic Floor Construction',
+                                      [wood, insulation, wood])
+    constr_set.floor_set.interior_construction = floor_constr
+    constr_set.roof_ceiling_set.exterior_construction = roof_constr
+    attic.properties.energy.construction_set = constr_set
+
+    Room.solve_adjacency([first_floor, second_floor, attic], 0.01)
+
+    model = Model('Multi Zone Single Family House', [first_floor, second_floor, attic])
+    model_dict = model.to_dict()
+
+    model_dict['properties']['energy']['program_types'][0] = office_program.to_dict()
+    model_dict['properties']['energy']['construction_sets'][0] = constr_set.to_dict()
+
+    rebuilt_model = Model.from_dict(model_dict)
+    assert rebuilt_model.rooms[0].properties.energy.program_type == office_program
+    assert rebuilt_model.rooms[2].properties.energy.construction_set == constr_set
+
+
 def test_writer_to_idf():
     """Test the Model to.idf method."""
     room = Room.from_box('Tiny House Zone', 5, 10, 3)
